+++ CHANGELOG.md
@@ -8,4 +8,5 @@ This changelog was created using the `clu` binary
 
 ### Features
 
+- (all) [#2](https://github.com/MalteHerrmann/contracts-testing/pull/2) Several minor additions for tests and scripts.
 - (ci) [#1](https://github.com/MalteHerrmann/contracts-testing/pull/1) Add changelog configuration and initialize empty changelog.